% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_derived_attributes.R
\name{em_derived_attributes}
\alias{em_derived_attributes}
\alias{em_da}
\title{Extended Metrics Dervied Columns}
\usage{
em_derived_attributes(df_string, col_string)
}
\arguments{
\item{df_string}{df name in string format}

\item{col_string}{col name in string format}
}
\value{
outputs the dataframe with new derived columns
}
\description{
input snow df & column to keyword search
}
\examples{
em_derived_attributes()
}
\seealso{
test test test
}
\author{
Chris
}
\keyword{calculated}
\keyword{metrics,}
