% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_keywords.R
\name{em_keywords}
\alias{em_keywords}
\title{input snow df & column to keyword search}
\usage{
em_keywords(df_string, col_string)
}
\arguments{
\item{df_string}{df name in string format}

\item{col_string}{col name in string format}
}
\description{
input snow df & column to keyword search
}
\examples{
em_keywords()
}
\keyword{metrics}
